requirements.txt
python-telegram-bot==20.7
google-auth==2.29.0
google-api-python-client==2.127.0
python-dotenv==1.0.1
pandas==2.2.2
